[DDP] Log unused param names under DETAIL debug mode.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/62209

When `TORCH_DISTRIBUTED_DEBUG=DETAIL` is set, log names and indices of unused parameters when searching for them.

Motivation is that we have seen a couple of issues occasionally when there are errors related to parameter possibly being marked as unused when it shouldn't, this can help narrow down the root cause by explicitly logging param names that are marked as unused.
ghstack-source-id: 134541461

Differential Revision: [D29916085](https://our.internmc.facebook.com/intern/diff/D29916085/)