[quant] update per-channel observer min/max_val attribute names
Summary:
This PR updates the attribute names from min_vals to min_val. the motivation for this is to keep the attribute name consistent with per-tensor observers so that dependencies (like FusedMovingAvgObsFakeQuantize) don't need to differentiate between the two observer types to access the attributes.

It also adds some BC tests to make sure that observers saved earlier with min_vals/max_vals can be loaded depending on the state_dict version.
Note: Scriptability of the observers isn't fully supported yet, so we aren't testing for that in this PR.

Test Plan:
python test/test_quantization.py TestSerialization

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 8f3bf2ed016bbe91dc91723a8d38c36ec0c38c66
Pull Request resolved: https://github.com/pytorch/pytorch/pull/62345